Update index.html
Do I need write access to commit a change? Text from alert from this page: "You’re making changes in a project you don’t have write access to. We’ve created a fork of this project for you to commit your proposed changes to. Submitting a change will write it to a new branch in your fork, so you can send a pull request."